build: Add cross-platform setup supporting linux and android
This is a massive commit. It introduces cross-platform configs to
support both linux and android. There are lots of new knobs which would
be documented in future commits.

+++ engine/src/2D/Actor.cxx
@@ -14,7 +14,7 @@ namespace RipsawEngine
 Actor::Actor(Engine* engine)
   : mEngine{engine}
 {
-  SDL_Log("[INFO] Actor (%ld bytes) created: %p", sizeof(*this), static_cast<void*>(this));
+  SDL_Log("[INFO] Actor (%zu bytes) created: %p", static_cast<size_t>(sizeof(*this)), static_cast<void*>(this));
   mEngine->addActor(this);
 }
